requirements.txt
Flask
supabase
python-dotenv
gunicorn
google-generativeai
openai
google-api-python-client
Flask-Session